Refractor the package name according to the specified format Update the Category and Transaction objects according to the swaggerhub API Add a Controller for the SessionID
