--- src/main/java/agilitytestbed/team/e/dpa/DpaApplication.java
@@ -1,12 +1,0 @@
-package agilitytestbed.team.e.dpa;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DpaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DpaApplication.class, args);
-	}
-}
