--- src/main/java/agilitytestbed/team/e/dpa/security/Authenticated.java
@@ -1,11 +1,0 @@
-package agilitytestbed.team.e.dpa.security;
-
-import javax.persistence.Entity;
-
-/**
- * @author Martijn Noorlander
- */
-@Entity
-public class Authenticated {
-
-}
