--- src/main/java/agilitytestbed/team/e/dpa/security/AuthenticatedService.java
@@ -1,16 +1,0 @@
-package agilitytestbed.team.e.dpa.security;
-
-import org.springframework.stereotype.Service;
-
-/**
- * @author Martijn Noorlander
- */
-
-@Service
-public interface AuthenticatedService {
-
-    public Authenticated getCurrent();
-
-    public void setCurrent(String sessionId);
-
-}
