--- src/main/java/agilitytestbed/team/e/dpa/transaction/NewTransactionDto.java
@@ -1,8 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-public class NewTransactionDto {
-}
