--- src/main/java/agilitytestbed/team/e/dpa/transaction/TransactionService.java
@@ -1,11 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction;
-
-import org.springframework.stereotype.Service;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-@Service
-public interface TransactionService {
-}
