--- src/main/java/agilitytestbed/team/e/dpa/transaction/TransactionServiceImpl.java
@@ -1,7 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction;
-
-/**
- * @author Martijn Noorlander
- */
-public class TransactionServiceImpl implements TransactionService {
-}
