--- src/main/java/agilitytestbed/team/e/dpa/transaction/category/Category.java
@@ -1,19 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction.category;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-@Entity
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-}
