--- src/main/java/agilitytestbed/team/e/dpa/transaction/category/CategoryRepository.java
@@ -1,10 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction.category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
