--- src/main/java/agilitytestbed/team/e/dpa/transaction/category/CategoryService.java
@@ -1,12 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction.category;
-
-import org.springframework.stereotype.Service;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-@Service
-public interface CategoryService {
-
-}
