--- src/main/java/agilitytestbed/team/e/dpa/transaction/category/CategoryServiceImpl.java
@@ -1,7 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction.category;
-
-/**
- * @author Martijn Noorlander
- */
-public class CategoryServiceImpl implements CategoryService {
-}
