--- src/main/java/agilitytestbed/team/e/dpa/transaction/category/NewCategoryDto.java
@@ -1,8 +1,0 @@
-package agilitytestbed.team.e.dpa.transaction.category;
-
-/**
- * @author Martijn Noorlander
- * @since 4-2-2018.
- */
-public class NewCategoryDto {
-}
