+++ src/main/java/nl/utwente/ing/team/e/dpa/DpaApplication.java
@@ -1,0 +1,12 @@
+package nl.utwente.ing.team.e.dpa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DpaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DpaApplication.class, args);
+	}
+}
