+++ src/main/java/nl/utwente/ing/team/e/dpa/security/Authenticated.java
@@ -1,0 +1,11 @@
+package nl.utwente.ing.team.e.dpa.security;
+
+import javax.persistence.Entity;
+
+/**
+ * @author Martijn Noorlander
+ */
+@Entity
+public class Authenticated {
+
+}
