+++ src/main/java/nl/utwente/ing/team/e/dpa/security/SessionControllerImpl.java
@@ -1,0 +1,15 @@
+package nl.utwente.ing.team.e.dpa.security;
+
+import org.springframework.http.ResponseEntity;
+
+/**
+ * @author Martijn Noorlander
+ * @since 28-02-2018
+ **/
+public class SessionControllerImpl implements SessionController {
+
+    @Override
+    public ResponseEntity<Object> getSession() {
+        return null;
+    }
+}
