+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/TransactionService.java
@@ -1,0 +1,11 @@
+package nl.utwente.ing.team.e.dpa.transaction;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * @author Martijn Noorlander
+ * @since 4-2-2018.
+ */
+@Service
+public interface TransactionService {
+}
