+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/TransactionServiceImpl.java
@@ -1,0 +1,7 @@
+package nl.utwente.ing.team.e.dpa.transaction;
+
+/**
+ * @author Martijn Noorlander
+ */
+public class TransactionServiceImpl implements TransactionService {
+}
