+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/TransactionType.java
@@ -1,0 +1,11 @@
+package nl.utwente.ing.team.e.dpa.transaction;
+
+/**
+ * @author Martijn Noorlander
+ * @since 28-02-2018
+ **/
+public enum TransactionType {
+
+    WITHDRAWAL, DEPOSIT;
+
+}
