+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/category/Category.java
@@ -1,0 +1,35 @@
+package nl.utwente.ing.team.e.dpa.transaction.category;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+/**
+ * @author Martijn Noorlander
+ * @since 4-2-2018.
+ */
+@Entity
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    private String name;
+
+    public Category() {
+    }
+
+    public Category(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
