+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/category/CategoryService.java
@@ -1,0 +1,12 @@
+package nl.utwente.ing.team.e.dpa.transaction.category;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * @author Martijn Noorlander
+ * @since 4-2-2018.
+ */
+@Service
+public interface CategoryService {
+
+}
