+++ src/main/java/nl/utwente/ing/team/e/dpa/transaction/category/CategoryServiceImpl.java
@@ -1,0 +1,7 @@
+package nl.utwente.ing.team.e.dpa.transaction.category;
+
+/**
+ * @author Martijn Noorlander
+ */
+public class CategoryServiceImpl implements CategoryService {
+}
