--- src/test/java/agilitytestbed/team/e/dpa/DpaApplicationTests.java
@@ -1,16 +1,0 @@
-package agilitytestbed.team.e.dpa;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DpaApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
