+++ src/test/java/nl/utwente/ing/team/e/DpaApplicationTests.java
@@ -1,0 +1,16 @@
+package nl.utwente.ing.team.e;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class DpaApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
